Move metric/binary prefixes and helper utilities into `tools.ts` and update exports/imports
Remove utils from public exports.
