+++ package.json
@@ -135,10 +135,6 @@
       "types": "./dist/tools.d.ts",
       "import": "./dist/tools.js"
     },
-    "./utils": {
-      "types": "./dist/utils.d.ts",
-      "import": "./dist/utils.js"
-    },
     "./warts": {
       "types": "./dist/warts.d.ts",
       "import": "./dist/warts.js"
