+++ src/expression.ts
@@ -1,11 +1,7 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 
-import {
-  ABSURD_EXPONENT,
-  BinaryPrefix,
-  MetricPrefix,
-  validateBigInt,
-} from './utils.js';
+import {ABSURD_EXPONENT, validateBigInt} from './utils.js';
+import {BinaryPrefix, MetricPrefix} from './tools.js';
 import {Pythagorean, AbsolutePitch} from './pythagorean.js';
 import {AbsoluteMosPitch, MosStep} from './diamond-mos.js';
 import {Fraction, lcm} from 'xen-dev-utils/fraction';
