+++ src/parser/expression.ts
@@ -50,17 +50,14 @@ import {
   binaryBroadcast,
   ternaryBroadcast,
 } from '../stdlib/index.js';
+import {ZERO, ONE, NEGATIVE_ONE, F} from '../utils.js';
 import {
   metricExponent,
-  ZERO,
-  ONE,
-  NEGATIVE_ONE,
-  hasOwn,
   binaryExponent,
   MetricPrefix,
   BinaryPrefix,
-  F,
-} from '../utils.js';
+  hasOwn,
+} from '../tools.js';
 import {
   pythagoreanMonzo,
   absoluteMonzo,
